chucha!

+++ app/Http/Controllers/Clases/ClaseController.php
@@ -2,112 +2,106 @@
 
 namespace App\Http\Controllers\Clases;
 
+use App\Http\Controllers\ApiController;
+use App\Models\Clases\Clase;
+use App\Models\Clases\Reservation;
+use App\Models\Clases\ClaseType;
+use App\Models\Plans\PlanUser;
 use Auth;
 use Carbon\Carbon;
-use App\Models\Clases\Clase;
 use Illuminate\Http\Request;
-use App\Models\Plans\PlanUser;
-use App\Models\Clases\Reservation;
-use App\Http\Controllers\ApiController;
 
 class ClaseController extends ApiController
 {
     /**
-     * Get all clases
-     * 
-     * @return Clase
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function store(Request $request)
     {
-        $clases = Clase::all();
+        //
+    }
 
+
+    public function index(Request $request)
+    {
+        $request->request->add(['sort_by_asc' => 'start','per_page' => 15]); //por ahora despues este request debe estar en la app cliente
+        $clases = Clase::all();
         return $this->showAll($clases);
     }
 
-    /**
-     * Get all days of the week
-     * 
-     * @return \Illuminate\Http\JsonResponse
-     */
-    public function week()
+    public function week(ClaseType $clase_type)
     {
+        
         $week = [];
-        $today = today();
+        $today = carbon::today();
         $date = $today;
         $day = [];
-        for ($i=0; $i < 7; $i++) {
+        for ($i = 0; $i <= 7; $i++) {
             $dow = $date->dayOfWeek;
-            $isToday = $i == 0 ? true : false;
+            if ($i == 0) {
+                $isToday = (bool) true;
+            } else {
+                $isToday = (bool) false;
+            }
 
-            $reservation_today = $this->hasClase($date);
-            $can_reserve = $this->canReserve($date);
-            $day = [
-                "date" => (string)$date->toDateString(),
-                "day" => (string)$date->format('d'),
-                "dayName" => (string)ucfirst($date->formatLocalized('%A' )),
+            $day_has_clases = $this->dayHasClases($date, $clase_type);
+            //$can_reserve = $this->canReserve($date);
+            $day = ["date" => (string) $date->toDateString(),
+                "day" => (string) $date->format('d'),
+                "dayName" => (string) ucfirst($date->formatLocalized('%A')),
                 "today" => $isToday,
-                "hasClases" => $reservation_today,
-                "canReserve" => $can_reserve,
+                "dayHasClases" => $day_has_clases,
+                "canReserve" => (bool) true,
+                "hasReserve" => (bool) false,
+                // "canReserve" => $can_reserve,
             ];
 
             $week = array_add($week, $dow, $day);
 
             $date = $date->addDay();
         }
-
+        $week[7] = $week[0];
         array_forget($week, '0');
 
         return response()->json(['data' => $week], 200);
+
     }
 
-    /**
-     * [historic description]
-     * 
-     * @return [type] [description]
-     */
     public function historic()
     {
-        $clases = Auth::user()->clases->where('date','<',today())
-                                      ->where(now()->format('H:i'), '<', 'finish_at');
+        $clases = Auth::user()->clases->where('date', '<', today())
+            ->where(now()->format('H:i'), '<', 'finish_at');
 
         return $this->showAll($clases);
     }
+    public function types()
+    {
+        $types = Clasetype::all();
 
-    /**
-     * [coming description]
-     * 
-     * @return [type] [description]
-     */
+        return  $this->showAll($types);
+    }
+
     public function coming()
     {
         $clases = Auth::user()->clases->where('date', '>=', today());
-                                      // ->where(now()->format('H:i'), '>=', 'finish_at');
+        // ->where(now()->format('H:i'), '>=', 'finish_at');
         return $this->showAll($clases);
     }
 
-    /**
-     * [users description]
-     * 
-     * @param  Clase  $clase [description]
-     * @return [type]        [description]
-     */
     public function users(Clase $clase)
     {
         $users = $clase->users;
-
+        //dd($users);
         return $this->showAll($users, 200);
     }
 
-    /**
-     * [reservations description]
-     * 
-     * @param  Clase  $clase [description]
-     * @return [type]        [description]
-     */
     public function reservations(Clase $clase)
     {
         $reservations = $clase->reservations;
-
+        //dd($users);
         return $this->showAll($reservations, 200);
     }
 
@@ -129,78 +123,74 @@ class ClaseController extends ApiController
      */
     public function reserve(Request $request, Clase $clase)
     {
-        $planuser = PlanUser::where('plan_status_id', 1)
-                            ->where('user_id', Auth::id())
-                            ->first();
-      $reservation = new Reservation;
-      $reservation->user_id = Auth::user()->id;
-      $reservation->clase_id = $clase->id;
-      $reservation->by_god = false;
-      $reservation->reservation_status_id = 1 ;
-      if($planuser){
-        $reservation->plan_user_id = $planuser->id;
-        if (!in_array($planuser->plan->id, $clase->block->plans->pluck('id')->toArray())) {
-          return $this->errorResponse('Tu plan no te deja tomar esta clase', 400);
+
+        $planuser = PlanUser::where('start_date', '<=', Carbon::parse($clase->date))
+            ->where('finish_date', '>=', Carbon::parse($clase->date))
+            ->where('user_id', Auth::id())
+            ->whereIn('plan_status_id', [1, 3])
+            ->first();
+        if (!$planuser) {
+            return $this->errorResponse('No tienes un plan que te permita tomar esta clase', 403);
         }
-      }
-      if($reservation->save())
-      {
-        $planuser->counter = $planuser->counter - 1;
-        $planuser->save();
-        return $this->showOne($reservation->clase, 200);
-      } else {
-        return $this->errorResponse('No se puede tomar la clase', 400);
-      }
+
+        if (count($clase->users) >= $clase->quota ) {
+            return $this->errorResponse('No puedes reservar, la clase esta llena.', 403);
+        }
+
+        $hasReserve = $this->hasReserve($clase);
+        if ($hasReserve) {
+            return $this->errorResponse($hasReserve, 403);
+        }
+        $reservation = new Reservation;
+        $reservation->user_id = Auth::user()->id;
+        $reservation->clase_id = $clase->id;
+        $reservation->by_god = false;
+        $reservation->reservation_status_id = 1;
+        $reservation->plan_user_id = $planuser->id;
+
+        if (!in_array($planuser->plan->id, $clase->block->plans->pluck('id')->toArray())) {
+            return $this->errorResponse('Tu plan no te deja tomar esta clase', 403);
+        }
+
+        if ($reservation->save()) {
+            $planuser->counter = $planuser->counter - 1;
+            $planuser->save();
+            return $this->showOne($reservation->clase, 201);
+        } else {
+            return $this->errorResponse('No se pudo tomar la clase', 400);
+        }
 
     }
 
-    /**
-     * [confirm description]
-     * 
-     * @param  Request $request [description]
-     * @param  Clase   $clase   [description]
-     * @return [type]           [description]
-     */
     public function confirm(Request $request, Clase $clase)
     {
-      $reservation = Reservation::where('clase_id', $clase->id)->where('user_id', Auth::id())->first();
-      if ($reservation == null) {
-          return $this->errorResponse('No puede confirmar una clase en la que no esta', 403);
-      }
-      $reservation->reservation_status_id = 2;
-      $reservation->save();
-      return $this->showOne($reservation->clase, 200);
+        $reservation = Reservation::where('clase_id', $clase->id)->where('user_id', Auth::id())->first();
+        if (!$reservation) {
+            return $this->errorResponse('No puede confirmar una clase en la que no esta', 403);
+        }
+        $reservation->reservation_status_id = 2;
+        $reservation->save();
+        return $this->showOne($reservation->clase, 201);
     }
 
-    /**
-     * [hasReserve description]
-     * 
-     * @param  [type]  $clase [description]
-     * @return boolean        [description]
-     */
     private function hasReserve($clase)
     {
         $response = '';
-        $clases = Clase::where('date', $clase->date)->get();
+        $clases = Clase::where('date', $clase->date)->where('clase_type_id', $clase->clase_type_id)->get();
         foreach ($clases as $clase) {
             $reservations = Reservation::where('user_id', Auth::id())->where('clase_id', $clase->id)->get();
             if (count($reservations) != 0) {
-                $response = 'Ya tiene clase tomada este dia';
+                $response = 'Ya tiene una clase tomada este dia';
             }
         }
         return $response;
     }
 
-    /**
-     * [hasClase description]
-     * 
-     * @param  DateTimeZone|string|null  $date 
-     * @return boolean       [description]
-     */
+
     private function hasClase($date)
     {
         $response = false;
-        $clases = Clase::where('date', $date)->get();
+        $clases = Clases::where('date', $date)->get();
         foreach ($clases as $clase) {
             $reservations = Reservation::where('user_id', Auth::id())->where('clase_id', $clase->id)->get();
             if (count($reservations) != 0) {
@@ -210,65 +200,94 @@ class ClaseController extends ApiController
         return $response;
     }
 
-    /**
-     * Check for the user not exceed the limit of day classes
-     * 
-     * @param  \DateTimeZone|string|null $date [description]
-     * @return boolean
-     */
+    private function dayHasClases($date, ClaseType $clase_type )
+    {
+        $response = false;
+        if(!$clase_type->exists){
+            $clase_type = ClaseType::first();
+        }
+
+
+        $clases = $clase_type->clases()->whereDate('date', $date->format('Y-m-d'))->first();
+        if($clases){
+            $response = true;
+        }
+
+        // if(!$response){
+        //     dd($date,$clase_type,$response, $clases);
+        // }
+        
+
+        return $response;
+    }
+
     public function canReserve($date)
     {
         $response = false;
-        $planusers = PlanUser::whereIn('plan_status_id', [1,3])
-                             ->where('user_id', Auth::id())
-                             ->get(['id', 'start_date', 'finish_date', 'counter']);
-
+        $planusers = PlanUser::whereIn('plan_status_id', [1, 3])->where('user_id', Auth::id())->get();
         foreach ($planusers as $planuser) {
+
             if ($date->between(Carbon::parse($planuser->start_date), Carbon::parse($planuser->finish_date))) {
                 if ($planuser->counter > 0) {
                     $response = true;
                 }
             }
-        }
 
+        }
+        
         return $response;
     }
 
-    /**
-     * [remove description]
-     * 
-     * @param  Request $request [description]
-     * @param  Clase   $clase   [description]
-     * @return [type]           [description]
-     */
     public function remove(Request $request, Clase $clase)
     {
         $reservation = Reservation::where('clase_id', $clase->id)->where('user_id', Auth::id())->first();
-        
         if ($reservation == null) {
             return $this->errorResponse('No puede votar una clase en la que no esta', 403);
         }
-        $planuser = PlanUser::where('plan_status_id', 1)->where('user_id', Auth::id())->first();
+        $planUser = Auth::user()->plan_users()->where('start_date', '<=', $clase->date)->where('finish_date', '>=', $clase->date) ->whereIn('plan_status_id', [1, 3])->first();
+
+        if(!$planUser){
+            return $this->errorResponse('no existe el plan', 403);
+        }
 
         if ($clase->date < toDay()->format('Y-m-d')) {
             return $this->errorResponse('No puede votar una clase de un día anterior a hoy', 403);
         } elseif ($clase->date > toDay()->format('Y-m-d')) {
             if ($reservation->delete()) {
-                    $planuser->counter = $planuser->counter + 1;
-                    $planuser->save();
-                    return $this->showOne($clase, 200);
-                }
+                $planUser->counter = $planUser->counter + 1;
+                $planUser->save();
+                return $this->showOne($clase, 200);
+            }
         } else {
             $class_hour = Carbon::parse($clase->start_at);
             if ($class_hour->diffInMinutes(now()->format('H:i')) < 40) {
                 return $this->errorResponse('Ya no puede votar la clase', 400);
             } else {
                 if ($reservation->delete()) {
-                    $planuser->counter = $planuser->counter - 1;
-                    $planuser->save();
+                    $planUser->counter = $planUser->counter - 1;
+                    $planUser->save();
                     return $this->showOne($clase, 200);
                 }
             }
         }
     }
+
 }
+// private function hasTwelvePlan($planuser)
+// {
+//     $responseTwo = null;
+//     if ($planuser->plan_id == 5 && $planuser->counter >= 12) {
+//         $responseTwo = 'No puede reservar, ya ha ocupado o reservado sus 12 clases del plan 12 clases mensual';
+//     }
+//     elseif ($planuser->plan_id == 6 && $planuser->counter >= 12) {
+//         $responseTwo = 'El plan de 12 clases trimestral no le permite tomar mas clases';
+//     }
+//     elseif ($planuser->plan_id == 7 && $planuser->counter >= 12) {
+//         $responseTwo = 'El plan de 12 clases semestral no le permite tomar mas clases';
+//     }
+//     elseif ($planuser->plan_id == 8 && $planuser->counter >= 12) {
+//         $responseTwo = 'El plan de 12 clases anual no le permite tomar mas clases';
+//     }
+
+//     return $responseTwo;
+// }
